refactor: merge local storage services
