--- src/client/services/index.ts
@@ -1,3 +1,0 @@
-export * from "./IndexedDatabaseService";
-export * from "./LocalStorageService";
-export * from "./StateStorageService";
