+++ src/services/LocalStorageService/index.ts
@@ -1,0 +1,1 @@
+export * from "./LocalStorageService.client";
